requirements.txt
Django==2.2.10
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
django-rest-swagger==2.1.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
coverage==4.5.2